Share code with migration prop test
